test(review): centralize env guard cleanup for review tests (#811)
