+++ crates/cli-sub-agent/src/review_consensus_tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::test_env_lock::ScopedEnvVarRestore;
 use csa_config::{GlobalConfig, ProjectMeta, ResourcesConfig, ToolConfig};
 use csa_core::env::{CSA_PARENT_SESSION_DIR_ENV_KEY, CSA_SESSION_DIR_ENV_KEY};
 use csa_session::review_artifact::{Finding, ReviewArtifact, Severity, SeveritySummary};
@@ -6,39 +7,6 @@ use std::sync::Mutex;
 use tempfile::tempdir;
 
 static REVIEWER_ENV_LOCK: Mutex<()> = Mutex::new(());
-
-struct ScopedEnvVar {
-    key: &'static str,
-    original: Option<String>,
-}
-
-impl ScopedEnvVar {
-    fn set(key: &'static str, value: &str) -> Self {
-        let original = std::env::var(key).ok();
-        // SAFETY: test-scoped env mutation guarded by REVIEWER_ENV_LOCK.
-        unsafe { std::env::set_var(key, value) };
-        Self { key, original }
-    }
-
-    fn unset(key: &'static str) -> Self {
-        let original = std::env::var(key).ok();
-        // SAFETY: test-scoped env mutation guarded by REVIEWER_ENV_LOCK.
-        unsafe { std::env::remove_var(key) };
-        Self { key, original }
-    }
-}
-
-impl Drop for ScopedEnvVar {
-    fn drop(&mut self) {
-        // SAFETY: test-scoped env mutation guarded by REVIEWER_ENV_LOCK.
-        unsafe {
-            match self.original.take() {
-                Some(value) => std::env::set_var(self.key, value),
-                None => std::env::remove_var(self.key),
-            }
-        }
-    }
-}
 
 fn project_config_with_enabled_tools(tools: &[&str]) -> ProjectConfig {
     let mut tool_map = HashMap::new();
@@ -267,8 +235,9 @@ fn build_multi_reviewer_instruction_prefers_current_session_dir_env_when_both_ex
     let _env_lock = REVIEWER_ENV_LOCK
         .lock()
         .expect("reviewer env lock poisoned");
-    let _parent_guard = ScopedEnvVar::set(CSA_PARENT_SESSION_DIR_ENV_KEY, "/tmp/parent-session");
-    let _session_guard = ScopedEnvVar::set(CSA_SESSION_DIR_ENV_KEY, "/tmp/child-session");
+    let _parent_guard =
+        ScopedEnvVarRestore::set(CSA_PARENT_SESSION_DIR_ENV_KEY, "/tmp/parent-session");
+    let _session_guard = ScopedEnvVarRestore::set(CSA_SESSION_DIR_ENV_KEY, "/tmp/child-session");
 
     let prompt = build_multi_reviewer_instruction("Base prompt", 2, ToolName::Codex, None);
 
@@ -284,8 +253,9 @@ fn build_multi_reviewer_instruction_falls_back_to_parent_session_dir_env() {
     let _env_lock = REVIEWER_ENV_LOCK
         .lock()
         .expect("reviewer env lock poisoned");
-    let _parent_guard = ScopedEnvVar::set(CSA_PARENT_SESSION_DIR_ENV_KEY, "/tmp/parent-session");
-    let _session_guard = ScopedEnvVar::unset(CSA_SESSION_DIR_ENV_KEY);
+    let _parent_guard =
+        ScopedEnvVarRestore::set(CSA_PARENT_SESSION_DIR_ENV_KEY, "/tmp/parent-session");
+    let _session_guard = ScopedEnvVarRestore::unset(CSA_SESSION_DIR_ENV_KEY);
 
     let prompt = build_multi_reviewer_instruction("Base prompt", 3, ToolName::Codex, None);
 
@@ -297,8 +267,8 @@ fn build_multi_reviewer_instruction_uses_session_first_shell_fallback_when_env_m
     let _env_lock = REVIEWER_ENV_LOCK
         .lock()
         .expect("reviewer env lock poisoned");
-    let _parent_guard = ScopedEnvVar::unset(CSA_PARENT_SESSION_DIR_ENV_KEY);
-    let _session_guard = ScopedEnvVar::unset(CSA_SESSION_DIR_ENV_KEY);
+    let _parent_guard = ScopedEnvVarRestore::unset(CSA_PARENT_SESSION_DIR_ENV_KEY);
+    let _session_guard = ScopedEnvVarRestore::unset(CSA_SESSION_DIR_ENV_KEY);
 
     let prompt = build_multi_reviewer_instruction("Base prompt", 4, ToolName::Codex, None);
